LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scroll-position-try-006-expected.txt


FAIL Should use the last fallback position initially assert_equals: Anchored element should be at the top of anchor expected 600 but got 100
FAIL Should use the third fallback position with enough space below assert_equals: Anchored element should be at the bottom of anchor expected 450 but got 0
FAIL Should use the second fallback position with enough space right assert_equals: Anchored element should be at the top of anchor expected 450 but got 100
FAIL Should use the first fallback position with enough space below and right assert_equals: Anchored element should be at the bottom of anchor expected 450 but got 0
